requirements.txt
mock==2.0.0
Flask_Env==1.0.1
Flask==0.12.2
azure_common==1.1.8
